+++ cli/src/django_ninja_matt/__init__.py
@@ -1,4 +1,4 @@
 """Django Ninja Matt - CLI tool for project scaffolding."""
 
-__version__ = "1.0.0"
+__version__ = "1.2.0"
 __author__ = "Matt Jaikaran"